requirements.txt
anyio==3.6.1
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
Naked==0.1.31
pycrypto==2.6.1
pydantic==1.10.2
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
